orders/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
billiard==3.6.4.0
black==21.6b0
brew==0.1.3
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
idna==3.2
kombu==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
pep8==1.7.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
ujson==4.0.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
